+++ ncvue/ncvutils.py
@@ -54,6 +54,7 @@ import tkinter as tk
 import numpy as np
 import matplotlib.dates as mpld
 import cartopy.crs as ccrs
+import platform
 
 
 __all__ = ['DIMMETHODS', 'SEPCHAR',
@@ -78,12 +79,54 @@ median - 50-percentile
 var - variance
 """
 
-
-SEPCHAR = chr(6)
+ios = platform.system()
+if ios == 'Darwin':
+    # SEPCHAR = chr(6)
+    SEPCHAR = chr(31)
+elif ios == 'Java':
+    SEPCHAR = chr(31)
+elif ios == 'Linux':
+    SEPCHAR = chr(31)
+elif ios == 'Windows':
+    SEPCHAR = chr(31)
+else:
+    SEPCHAR = chr(31)
 """
 Invisible character to split strings.
 
-ASCII character 6 = ACKNOWLEDGE (ACK)
+Name (Code) Decimal
+Null (NUL) 0
+Start of Heading (SOH) 1
+Start of Text (STX) 2
+End of Text (ETX) 3
+End of Transmit (EOT) 4
+Enquiry (ENQ) 5
+Acknowledge (ACK) 6
+Bell (BEL) 7
+Back Space (BS) 8
+Horizontal Tab (TAB) 9
+Line Feed (LF) 10
+Vertical Tab (VT) 11
+Form Feed (FF) 12
+Carriage Return (CR) 13
+Shift Out (SO) 14
+Shift In (SI) 15
+Data Line Escape (DLE) 16
+Device Control 1 (DC1) 17
+Device Control 2 (DC2) 18
+Device Control 3 (DC3) 19
+Device Control 4 (DC4) 20
+Negative Acknowledge (NAK) 21
+Synchronous Idle (SYN) 22
+End of Transmit Block (ETB) 23
+Cancel (CAN) 24
+End of Medium (EM) 25
+Substitute (SUB) 26
+Escape (ESC) 27
+File Separator (FS) 28
+Group Separator (GS) 29
+Record Separator (RS) 30
+Unit Separator (US) 31
 """